Got it working initially
Using SEGGER debugger firmware on my ST-Link. Was getting hard faults using ST-Link firmware (debugging application code using external SDRAM as main memory, not internal SRAM; only got the hard fault in the debugging scenario -- pressing reset on the device after the application was loaded to the QSPI FLASH let the application run fine, same with debugging the bootloader that configures the QSPI FLASH and SDRAM and then jumps to the application in QSPI FLASH). Kind of an anti-climactic ending to what's been a thorn in my side for over a week since I don't know exactly what was wrong, I only know scenarios that don't work and things that are definitely not wrong. I think this has to do with how SEGGER sets the program counter in the application code. Comforted by all the options in the debug configurations vs what was there for ST-Link.